Split timing functions from simif.h
